+++ lib/active_record/connection_adapters/elasticsearch/schema_definitions/column_methods.rb
@@ -14,12 +14,12 @@ module ActiveRecord
                                 :long, :integer, :short, :byte, :double, :float, :half_float, :scaled_float, :unsigned_long,
                                 :date, :date_nanos,
                                 :alias,
-                                :object, :flattened, :nested, :join,
+                                :object, :flattened, :nested, :join, :passthrough,
                                 :integer_range, :float_range, :long_range, :double_range, :date_range, :ip_range,
-                                :ip, :version, :murmur3,
+                                :ip, :version, :murmur3, :counted_keyword,
                                 :aggregate_metric_double, :histogram,
                                 :text, :match_only_text, :completion, :search_as_you_type, :token_count, :semantic_text,
-                                :dense_vector, :sparse_vector, :rank_feature, :rank_features,
+                                :dense_vector, :sparse_vector, :rank_feature, :rank_features, :rank_vectors,
                                 :geo_point, :geo_shape, :point, :shape,
                                 :percolator